requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
anthropic==0.34.2
python-dotenv==1.0.1
